Move code into one file to keep processor logic together
